% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bec.R
\name{bec}
\alias{bec}
\alias{dbec}
\alias{sbec}
\alias{hbec}
\alias{rbec}
\title{Bivariate Extended Chen distribution: density, survival, hazard and
random generation}
\usage{
dbec(x1, x2, alpha1, alpha2, alpha3, beta, lambda, log = FALSE)

sbec(x1, x2, alpha1, alpha2, alpha3, beta, lambda)

hbec(x1, x2, alpha1, alpha2, alpha3, beta, lambda)

rbec(n, alpha1, alpha2, alpha3, beta, lambda)
}
\arguments{
\item{x1, x2}{numeric vectors of first and second component.}

\item{alpha1, alpha2, alpha3, beta, lambda}{positive parameters.}

\item{log}{logical; if \code{TRUE} the log density is returned.}

\item{n}{number of observations.}
}
\value{
A numeric vector, or for \code{rbec} a two column matrix with
  columns \code{x1} and \code{x2}.
}
\description{
Probability density, survival function, hazard function and random
generation for the Marshall-Olkin type Bivariate Extended Chen
distribution \eqn{BEC(\alpha_1,\alpha_2,\alpha_3,\beta,\lambda)}.
}
\details{
With \eqn{U_i \sim EC(\alpha_i,\beta,\lambda)} independent and
  \eqn{X_1=\min(U_1,U_3)}, \eqn{X_2=\min(U_2,U_3)}, the joint survival
  function is
  \deqn{S(x_1,x_2)=\begin{cases}
    S_{EC}(x_1;\alpha_1,\beta,\lambda)\,
      S_{EC}(x_2;\alpha_2+\alpha_3,\beta,\lambda) & x_2>x_1\\
    S_{EC}(x_1;\alpha_1+\alpha_3,\beta,\lambda)\,
      S_{EC}(x_2;\alpha_2,\beta,\lambda) & x_1>x_2\\
    S_{EC}(x;\alpha_1+\alpha_2+\alpha_3,\beta,\lambda) & x_1=x_2=x,\end{cases}}
  where \eqn{S_{EC}(t;\alpha,\beta,\lambda)=[1+\lambda(e^{t^\beta}-1)]^{-\alpha}}
  (see \code{\link{sech}}). The joint density is
  \deqn{f(x_1,x_2)=\begin{cases} f_1 & x_2>x_1\\ f_2 & x_1>x_2\\
    f_3 & x_1=x_2=x,\end{cases}}
  with
  \deqn{f_1=\alpha_1(\alpha_2+\alpha_3)\lambda^2\beta^2
    x_1^{\beta-1}x_2^{\beta-1}e^{x_1^{\beta}+x_2^{\beta}}
    [1+\lambda(e^{x_1^{\beta}}-1)]^{-(\alpha_1+1)}
    [1+\lambda(e^{x_2^{\beta}}-1)]^{-(\alpha_2+\alpha_3+1)},}
  \deqn{f_2=\alpha_2(\alpha_1+\alpha_3)\lambda^2\beta^2
    x_1^{\beta-1}x_2^{\beta-1}e^{x_1^{\beta}+x_2^{\beta}}
    [1+\lambda(e^{x_1^{\beta}}-1)]^{-(\alpha_1+\alpha_3+1)}
    [1+\lambda(e^{x_2^{\beta}}-1)]^{-(\alpha_2+1)},}
  \deqn{f_3(x)=\alpha_3\beta\lambda x^{\beta-1}e^{x^{\beta}}
    [1+\lambda(e^{x^{\beta}}-1)]^{-(\alpha_1+\alpha_2+\alpha_3+1)}.}
  The three regions carry mass \eqn{\alpha_i/(\alpha_1+\alpha_2+\alpha_3)},
  so the density integrates to one. The hazard function is \eqn{f/S},
  \deqn{h(x_1,x_2)=\begin{cases}
    \dfrac{\alpha_1(\alpha_2+\alpha_3)\lambda^2\beta^2x_1^{\beta-1}
      x_2^{\beta-1}e^{x_1^{\beta}+x_2^{\beta}}}
      {[1+\lambda(e^{x_1^{\beta}}-1)][1+\lambda(e^{x_2^{\beta}}-1)]}
      & x_2>x_1\\[2ex]
    \dfrac{\alpha_2(\alpha_1+\alpha_3)\lambda^2\beta^2x_1^{\beta-1}
      x_2^{\beta-1}e^{x_1^{\beta}+x_2^{\beta}}}
      {[1+\lambda(e^{x_1^{\beta}}-1)][1+\lambda(e^{x_2^{\beta}}-1)]}
      & x_1>x_2\\[2ex]
    \dfrac{\alpha_3\beta\lambda x^{\beta-1}e^{x^{\beta}}}
      {1+\lambda(e^{x^{\beta}}-1)} & x_1=x_2=x.\end{cases}}
  The componentwise hazard gradients are available separately through
  \code{\link{hbec1}} and \code{\link{hbec2}}.
}
\examples{
dbec(0.5, 0.9, alpha1 = 1.5, alpha2 = 2.5, alpha3 = 0.5, beta = 1.8, lambda = 2)
sbec(0.5, 0.9, alpha1 = 1.5, alpha2 = 2.5, alpha3 = 0.5, beta = 1.8, lambda = 2)
hbec(0.5, 0.9, alpha1 = 1.5, alpha2 = 2.5, alpha3 = 0.5, beta = 1.8, lambda = 2)
set.seed(1)
rbec(5, alpha1 = 1.5, alpha2 = 2.5, alpha3 = 0.5, beta = 1.8, lambda = 2)

}
\seealso{
\code{\link{sbec}}, \code{\link{hbec}}, \code{\link{rbec}},
  \code{\link{fitbec}}, \code{\link{sech}}
}
