% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BvCh-package.R
\docType{package}
\name{BvCh-package}
\alias{BvCh-package}
\alias{BvCh}
\title{Bivariate Chen Distribution: Density, Survival, Hazard, Plots and Estimates}
\description{
Implements probability density, survival and hazard functions, random
generation, graphical displays and maximum likelihood estimation for a
family of bivariate Chen type lifetime models.
}
\section{Models}{

\describe{
  \item{Bivariate Chen (BvCh)}{Marshall-Olkin type model of Gupta, Kumar and
    co-authors. With \eqn{U_i \sim Chen(\alpha_i, \beta)} independent and
    \eqn{Z_1 = \min(U_1, U_3)}, \eqn{Z_2 = \min(U_2, U_3)}, the pair
    \eqn{(Z_1,Z_2)} is \eqn{BvCh(\alpha_1,\alpha_2,\alpha_3,\beta)}. The
    distribution is a mixture of an absolutely continuous part of total
    mass \eqn{(\alpha_1+\alpha_2)/(\alpha_1+\alpha_2+\alpha_3)} and a
    singular part of mass \eqn{\alpha_3/(\alpha_1+\alpha_2+\alpha_3)}
    carried on the diagonal.}
  \item{Bivariate Extended Chen (BEC)}{Marshall-Olkin type model using
    extended Chen latent variables, adding a shape parameter
    \eqn{\lambda > 0}.}
  \item{Bivariate Inverse FGM Chen (IFGM-Chen)}{Copula construction pairing
    the Inverse FGM copula with Chen marginals. This model is absolutely
    continuous and has no singular component.}
}
}

\section{Univariate Chen}{

All univariate Chen building blocks are taken from the \pkg{reliaR}
package rather than being re-implemented. In particular
\code{reliaR::schen(x, beta, lambda)} satisfies
\code{schen(x, beta, lambda) == exp(lambda * (1 - exp(x^beta))}, which is
the \eqn{S_{Ch}(z;\alpha,\beta)} used throughout the bivariate formulae.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/puneetstat999/BvCh}
  \item Report bugs at \url{https://github.com/puneetstat999/BvCh/issues}
}

}
\author{
\strong{Maintainer}: Puneet Kumar Gupta \email{puneetstat999@gmail.com}

Authors:
\itemize{
  \item Puneet Kumar Gupta \email{puneetstat999@gmail.com}
  \item Mukul Bijalwan (Orchid ID: https://orcid.org/0009-0001-3040-6912)
}

}
\keyword{internal}
