% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvch.R
\name{bvch_components}
\alias{bvch_components}
\alias{dbvch_ac}
\alias{dbvch_sing}
\title{Absolutely continuous and singular components of the BvCh density}
\usage{
dbvch_ac(z1, z2, alpha1, alpha2, alpha3, beta)

dbvch_sing(z1, z2, alpha1, alpha2, alpha3, beta)
}
\arguments{
\item{z1, z2}{numeric vectors of first and second component.}

\item{alpha1, alpha2, alpha3, beta}{positive parameters.}
}
\value{
A numeric vector.
}
\description{
Decomposition of the Bivariate Chen density into the two components of the
mixture representation
\eqn{f = \frac{\alpha_1+\alpha_2}{\alpha_1+\alpha_2+\alpha_3}f_a +
\frac{\alpha_3}{\alpha_1+\alpha_2+\alpha_3}f_s}.
}
\details{
\code{dbvch_ac} is the properly normalised absolutely continuous
  component of Proposition 1; it integrates to one over the plane, and
  \code{dbvch_sing} integrates to one along the diagonal. Consequently
  \code{dbvch} is recovered exactly as
  \code{(alpha1 + alpha2)/(alpha1 + alpha2 + alpha3) * dbvch_ac(...)} plus
  \code{alpha3/(alpha1 + alpha2 + alpha3) * dbvch_sing(...)}.
}
\examples{
a <- 1.5; b2 <- 2.5; c3 <- 0.5; bt <- 1.8
zz1 <- c(0.4, 0.9, 0.7); zz2 <- c(0.6, 0.4, 0.7)
ac <- dbvch_ac(zz1, zz2, a, b2, c3, bt)
sg <- dbvch_sing(zz1, zz2, a, b2, c3, bt)
k <- (a + b2 + c3)
mix <- (a + b2) / k * ac + c3 / k * sg
all.equal(mix, dbvch(zz1, zz2, a, b2, c3, bt))

}
\seealso{
\code{\link{dbvch}}
}
