% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bec.R
\name{hbec_grad}
\alias{hbec_grad}
\alias{hbec1}
\alias{hbec2}
\title{Componentwise hazard gradients of the BEC model}
\usage{
hbec1(x1, x2, alpha1, alpha2, alpha3, beta, lambda)

hbec2(x1, x2, alpha1, alpha2, alpha3, beta, lambda)
}
\arguments{
\item{x1, x2}{numeric vectors of first and second component.}

\item{alpha1, alpha2, alpha3, beta, lambda}{positive parameters.}
}
\value{
A numeric vector.
}
\description{
The hazard gradients of the Bivariate Extended Chen model, that is the
hazard of the \eqn{i}-th marginal conditional on the other component.
}
\details{
\deqn{h_{X_1}(x_1,x_2)=\begin{cases}
  \dfrac{\alpha_1\beta\lambda x_1^{\beta-1}e^{x_1^{\beta}}}
    {1+\lambda(e^{x_1^{\beta}}-1)} & x_2>x_1\\
  \dfrac{(\alpha_1+\alpha_3)\beta\lambda x_1^{\beta-1}e^{x_1^{\beta}}}
    {1+\lambda(e^{x_1^{\beta}}-1)} & x_1>x_2\\
  \dfrac{(\alpha_1+\alpha_2+\alpha_3)\beta\lambda x_1^{\beta-1}
    e^{x_1^{\beta}}}{1+\lambda(e^{x_1^{\beta}}-1)} & x_1=x_2,\end{cases}}
with \eqn{h_{X_2}} obtained by interchanging the roles of \eqn{x_1},
\eqn{x_2} and \eqn{\alpha_1}, \eqn{\alpha_2}.
}
\examples{
hbec1(0.5, 0.9, alpha1 = 1.5, alpha2 = 2.5, alpha3 = 0.5, beta = 1.8, lambda = 2)
hbec2(0.5, 0.9, alpha1 = 1.5, alpha2 = 2.5, alpha3 = 0.5, beta = 1.8, lambda = 2)

}
\seealso{
\code{\link{hbec}}
}
